+++ packages/storage/sqlite-node/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added bounded active-branch queries to the SQLite session reader.
+
 ## [0.83.0] - 2026-07-29
 
 ## [0.82.1] - 2026-07-25